+++ src/charging-station/ocpp/auth/utils/AuthConfigValidator.ts
